move owneinfo
