+++ ydb/core/persqueue/pqtablet/partition/ownerinfo.h
